캐릭터 애니메이터 구조 변경
