+++ MAM/Assets/Animations/Characters/Bubble.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0d541a13f4925b64a96a67e89f9b1428
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
