+++ MAM/Assets/Animations/Characters/Bubble/BaseCharBubble.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0a225520b8ade84d91c5c1ac0b40439
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
