+++ MAM/Assets/Animations/Characters/Bubble/Bubble_Floating.anim.meta
