+++ MAM/Assets/Animations/Characters/Bubble/Bubble_Inactive.anim.meta
