+++ MAM/Assets/Animations/Characters/Bubble/Emote_Angry.anim.meta
