+++ MAM/Assets/Animations/Characters/Bubble/Emote_ExclamationBlue.anim
