+++ MAM/Assets/Animations/Characters/Bubble/Emote_ExclamationRed.anim
