+++ MAM/Assets/Animations/Characters/Bubble/Emote_Shy.anim.meta
