+++ MAM/Assets/Animations/Characters/Bubble/Emote_ThumbUpLeft.anim.meta
