+++ MAM/Assets/Animations/Characters/Bubble/Emote_ThumbUpRight.anim.meta
