+++ MAM/Assets/Animations/Characters/Bubble/Emote_Water.anim.meta
