+++ MAM/Assets/Animations/Characters/Bubble/Talk_DotDotDot.anim
