+++ MAM/Assets/Animations/Characters/Bubble/Talk_Idle.anim
