+++ MAM/Assets/Prefabs/Common/CharacterBase.prefab
@@ -209,6 +209,7 @@ Transform:
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 4498326202830219771}
+  - {fileID: 3230047090217657644}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &6993091677450593253
@@ -227,6 +228,7 @@ MonoBehaviour:
   _detector: {fileID: 3725084601284872886}
   _spriteResolver: {fileID: 9123878462812166935}
   _animator: {fileID: 6241298353178126811}
+  _speechBubbleAnimator: {fileID: 4027941042075318886}
 --- !u!1 &2888775006157134406
 GameObject:
   m_ObjectHideFlags: 0
@@ -306,8 +308,7 @@ Transform:
   m_LocalPosition: {x: 0, y: 1.04, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
-  m_Children:
-  - {fileID: 6319435961191248806}
+  m_Children: []
   m_Father: {fileID: 4498326202830219771}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!212 &1795721935061058264
@@ -430,6 +431,88 @@ MonoBehaviour:
   m_labelHash: 0.00028302343
   m_SpriteKey: 0
   m_SpriteHash: 1011805508
+--- !u!1 &4880591726492456252
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 3230047090217657644}
+  - component: {fileID: 1537290065166243912}
+  - component: {fileID: 7884082722003531698}
+  - component: {fileID: 4027941042075318886}
+  m_Layer: 0
+  m_Name: SpeechBubbleGroup
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &3230047090217657644
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4880591726492456252}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 1.03, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children:
+  - {fileID: 6319435961191248806}
+  m_Father: {fileID: 5880637469130732476}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!210 &1537290065166243912
+SortingGroup:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4880591726492456252}
+  m_Enabled: 1
+  m_SortingLayerID: -1491952087
+  m_SortingLayer: 2
+  m_SortingOrder: 1
+  m_SortAtRoot: 0
+--- !u!95 &7884082722003531698
+Animator:
+  serializedVersion: 7
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4880591726492456252}
+  m_Enabled: 1
+  m_Avatar: {fileID: 0}
+  m_Controller: {fileID: 9100000, guid: c0a225520b8ade84d91c5c1ac0b40439, type: 2}
+  m_CullingMode: 0
+  m_UpdateMode: 0
+  m_ApplyRootMotion: 0
+  m_LinearVelocityBlending: 0
+  m_StabilizeFeet: 0
+  m_AnimatePhysics: 0
+  m_WarningMessage: 
+  m_HasTransformHierarchy: 1
+  m_AllowConstantClipSamplingOptimization: 1
+  m_KeepAnimatorStateOnDisable: 0
+  m_WriteDefaultValuesOnDisable: 0
+--- !u!114 &4027941042075318886
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4880591726492456252}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f7673551028e94541b63677ef7c9af04, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _animator: {fileID: 7884082722003531698}
 --- !u!1 &9008355678983590792
 GameObject:
   m_ObjectHideFlags: 0
@@ -439,7 +522,6 @@ GameObject:
   serializedVersion: 6
   m_Component:
   - component: {fileID: 6319435961191248806}
-  - component: {fileID: 3003803701817534141}
   m_Layer: 0
   m_Name: SpeechBubble
   m_TagString: Untagged
@@ -456,22 +538,10 @@ Transform:
   m_GameObject: {fileID: 9008355678983590792}
   serializedVersion: 2
   m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
-  m_LocalPosition: {x: 0, y: 0.37400007, z: 0}
+  m_LocalPosition: {x: 0, y: 0.374, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
   - {fileID: 5250076623690867041}
-  m_Father: {fileID: 8354655507307387443}
+  m_Father: {fileID: 3230047090217657644}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!210 &3003803701817534141
-SortingGroup:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 9008355678983590792}
-  m_Enabled: 1
-  m_SortingLayerID: -1491952087
-  m_SortingLayer: 2
-  m_SortingOrder: 1
-  m_SortAtRoot: 0
