+++ MAM/Assets/Scripts/ManagerScene/Student/StudentCharacter/SpeechBubbleAnimator.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f7673551028e94541b63677ef7c9af04
