+++ MAM/Assets/Scripts/ManagerScene/Student/StudentCharacter/StudentCharacter.cs
@@ -7,11 +7,13 @@ public class StudentCharacter : MonoBehaviour
     [SerializeField] private CharacterClickDetector _detector = null;
     [SerializeField] private SpriteResolver _spriteResolver = null;
     [SerializeField] private CharacterAnimator _animator = null;
+    [SerializeField] private SpeechBubbleAnimator _speechBubbleAnimator = null;
     
     private readonly string _startCategory = "Idle_Front";
     private readonly string _startLabel = "Idle_Front_0";
     
     public CharacterAnimator Animator => _animator;
+    public SpeechBubbleAnimator SpeechBubbleAnimator => _speechBubbleAnimator;
     public string ID {get; set;}
     
     public void SetSpriteLibrary(SpriteLibraryAsset asset)
